feat: update payment and subscription models to use integer for price and amount, and set default statuses

+++ src/infra/db/models/plan.py
@@ -8,5 +8,5 @@ from src.infra.db.models.mixins import CreatedAtMixin, IdMixin, UpdatedAtMixin
 
 class Plan(Base, IdMixin, CreatedAtMixin, UpdatedAtMixin):
     name: Mapped[str] = mapped_column(unique=True, nullable=False)
-    price_usd: Mapped[float] = mapped_column(nullable=False)
+    price_usd: Mapped[int] = mapped_column(nullable=False)  # in cents
     duration_days: Mapped[int] = mapped_column(nullable=False)
